Generates the optimum expressed regions

Generates the optimum expressed regions

ODER_strand(
  bw_pos,
  bw_neg,
  auc_raw_pos,
  auc_raw_neg,
  auc_target,
  chrs = "",
  genome = "hg38",
  mccs,
  mrgs,
  gtf = NULL,
  ucsc_chr,
  ignore.strand,
  exons_no_overlap = NULL,
  bw_chr = "chr"
)

Arguments

bw_pos
positive strand bigwig file

bw_neg
negative strand bigwig file

auc_raw_pos
vector containing AUCs(Area Under Coverage) matching the order of bigwig paths.

auc_raw_neg
vector containing AUCs(Area Under Coverage) matching the order of bigwig paths.

auc_target
total AUC to normalise all samples to e.g. 40e6 * 100 would be the estimated total auc for sample sequenced to 40 million reads of 100bp in length.

chrs
chromosomes to obtain mean coverage for, default is "" giving every chromosome. Can take UCSC format(chrs = "chr1") or just the chromosome i.e. chrs = c(1,X)

genome
the UCSC genome you want to use, the default is hg38.

mccs
mean coverage cut-offs to apply.

mrgs
max region gaps to apply.

gtf
Either a string containg the path to a .gtf file or a pre-imported gtf using import.

ucsc_chr
logical scalar, determining whether to add "chr" prefix to the seqnames of non-overlapping exons and change "chrMT" -> "chrM". Note, if set to TRUE and seqnames already have "chr", it will not add another.

ignore.strand
logical value for input into findOverlaps, default is True.

exons_no_overlap
Optimum set of exons to help calculate deltas

bw_chr
specifies whether the bigwig files has the chromosomes labelled with a "chr" preceding the chromosome i.e. "chr1" vs "1". Can be either "chr" or "nochr" with "chr" being the default.

Value

list containing optimised ERs, optimal pair of MCC/MRGs and delta_df

Examples

library("magrittr")
gtex_metadata <- recount::all_metadata("gtex")

#> 2021-08-24 21:48:46 downloading the metadata to /tmp/RtmpfgXOQD/metadata_clean_gtex.Rdata

gtex_metadata <- gtex_metadata %>%
    as.data.frame() %>%
    dplyr::filter(project == "SRP012682")

gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)

url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)
bw_plus <- ODER:::.file_cache(url[58])
bw_minus <- ODER:::.file_cache(url[84])

opt_ers <- ODER_strand(
    bw_pos = bw_plus, bw_neg = bw_minus,
    auc_raw_pos = gtex_metadata[["auc"]][58], auc_raw_neg = gtex_metadata[["auc"]][84],
    auc_target = 40e6 * 100, chrs = "", genome = "hg38",
    mccs = c(5, 10), mrgs = c(10, 20), gtf = gtf_path, ucsc_chr = TRUE,
    ignore.strand = FALSE, exons_no_overlap = NULL, bw_chr = "chr"
)

#> [1] "2021-08-24 21:48:48 - Obtaining mean coverage across 1 samples"
#> [1] "2021-08-24 21:48:48 - chr1"
#> [1] "2021-08-24 21:48:51 - chr2"
#> [1] "2021-08-24 21:48:53 - chr3"
#> [1] "2021-08-24 21:48:54 - chr4"
#> [1] "2021-08-24 21:48:55 - chr5"
#> [1] "2021-08-24 21:48:56 - chr6"
#> [1] "2021-08-24 21:48:57 - chr7"
#> [1] "2021-08-24 21:48:58 - chr8"
#> [1] "2021-08-24 21:48:59 - chr9"
#> [1] "2021-08-24 21:49:00 - chr10"
#> [1] "2021-08-24 21:49:01 - chr11"
#> [1] "2021-08-24 21:49:02 - chr12"
#> [1] "2021-08-24 21:49:03 - chr13"
#> [1] "2021-08-24 21:49:03 - chr14"
#> [1] "2021-08-24 21:49:04 - chr15"
#> [1] "2021-08-24 21:49:04 - chr16"
#> [1] "2021-08-24 21:49:05 - chr17"
#> [1] "2021-08-24 21:49:06 - chr18"
#> [1] "2021-08-24 21:49:07 - chr19"
#> [1] "2021-08-24 21:49:08 - chr20"
#> [1] "2021-08-24 21:49:08 - chr21"
#> [1] "2021-08-24 21:49:09 - chr22"
#> [1] "2021-08-24 21:49:09 - chrX"
#> [1] "2021-08-24 21:49:10 - chrY"
#> [1] "2021-08-24 21:49:10 - chrM"
#> [1] "2021-08-24 21:49:10 - Obtaining mean coverage across 1 samples"
#> [1] "2021-08-24 21:49:10 - chr1"
#> [1] "2021-08-24 21:49:12 - chr2"
#> [1] "2021-08-24 21:49:13 - chr3"
#> [1] "2021-08-24 21:49:16 - chr4"
#> [1] "2021-08-24 21:49:16 - chr5"
#> [1] "2021-08-24 21:49:17 - chr6"
#> [1] "2021-08-24 21:49:18 - chr7"
#> [1] "2021-08-24 21:49:19 - chr8"
#> [1] "2021-08-24 21:49:20 - chr9"
#> [1] "2021-08-24 21:49:21 - chr10"
#> [1] "2021-08-24 21:49:22 - chr11"
#> [1] "2021-08-24 21:49:23 - chr12"
#> [1] "2021-08-24 21:49:25 - chr13"
#> [1] "2021-08-24 21:49:25 - chr14"
#> [1] "2021-08-24 21:49:26 - chr15"
#> [1] "2021-08-24 21:49:27 - chr16"
#> [1] "2021-08-24 21:49:27 - chr17"
#> [1] "2021-08-24 21:49:29 - chr18"
#> [1] "2021-08-24 21:49:30 - chr19"
#> [1] "2021-08-24 21:49:31 - chr20"
#> [1] "2021-08-24 21:49:31 - chr21"
#> [1] "2021-08-24 21:49:32 - chr22"
#> [1] "2021-08-24 21:49:32 - chrX"
#> [1] "2021-08-24 21:49:33 - chrY"
#> [1] "2021-08-24 21:49:33 - chrM"
#> [1] "2021-08-24 21:49:33 - Generating ERs for chr1"
#> [1] "2021-08-24 21:49:38 - Generating ERs for chr2"
#> [1] "2021-08-24 21:49:43 - Generating ERs for chr3"
#> [1] "2021-08-24 21:49:45 - Generating ERs for chr4"
#> [1] "2021-08-24 21:49:47 - Generating ERs for chr5"
#> [1] "2021-08-24 21:49:49 - Generating ERs for chr6"
#> [1] "2021-08-24 21:49:51 - Generating ERs for chr7"
#> [1] "2021-08-24 21:49:52 - Generating ERs for chr8"
#> [1] "2021-08-24 21:49:55 - Generating ERs for chr9"
#> [1] "2021-08-24 21:49:56 - Generating ERs for chr10"
#> [1] "2021-08-24 21:49:58 - Generating ERs for chr11"
#> [1] "2021-08-24 21:49:59 - Generating ERs for chr12"
#> [1] "2021-08-24 21:50:01 - Generating ERs for chr13"
#> [1] "2021-08-24 21:50:02 - Generating ERs for chr14"
#> [1] "2021-08-24 21:50:03 - Generating ERs for chr15"
#> [1] "2021-08-24 21:50:05 - Generating ERs for chr16"
#> [1] "2021-08-24 21:50:06 - Generating ERs for chr17"
#> [1] "2021-08-24 21:50:08 - Generating ERs for chr18"
#> [1] "2021-08-24 21:50:09 - Generating ERs for chr19"
#> [1] "2021-08-24 21:50:11 - Generating ERs for chr20"
#> [1] "2021-08-24 21:50:12 - Generating ERs for chr21"
#> [1] "2021-08-24 21:50:12 - Generating ERs for chr22"
#> [1] "2021-08-24 21:50:13 - Generating ERs for chrX"
#> [1] "2021-08-24 21:50:14 - Generating ERs for chrY"
#> [1] "2021-08-24 21:50:15 - Generating ERs for chrM"
#> [1] "2021-08-24 21:50:16 - Generating ERs for chr1"
#> [1] "2021-08-24 21:50:19 - Generating ERs for chr2"
#> [1] "2021-08-24 21:50:22 - Generating ERs for chr3"
#> [1] "2021-08-24 21:50:24 - Generating ERs for chr4"
#> [1] "2021-08-24 21:50:27 - Generating ERs for chr5"
#> [1] "2021-08-24 21:50:29 - Generating ERs for chr6"
#> [1] "2021-08-24 21:50:30 - Generating ERs for chr7"
#> [1] "2021-08-24 21:50:33 - Generating ERs for chr8"
#> [1] "2021-08-24 21:50:34 - Generating ERs for chr9"
#> [1] "2021-08-24 21:50:35 - Generating ERs for chr10"
#> [1] "2021-08-24 21:50:37 - Generating ERs for chr11"
#> [1] "2021-08-24 21:50:38 - Generating ERs for chr12"
#> [1] "2021-08-24 21:50:41 - Generating ERs for chr13"
#> [1] "2021-08-24 21:50:42 - Generating ERs for chr14"
#> [1] "2021-08-24 21:50:43 - Generating ERs for chr15"
#> [1] "2021-08-24 21:50:44 - Generating ERs for chr16"
#> [1] "2021-08-24 21:50:45 - Generating ERs for chr17"
#> [1] "2021-08-24 21:50:47 - Generating ERs for chr18"
#> [1] "2021-08-24 21:50:47 - Generating ERs for chr19"
#> [1] "2021-08-24 21:50:49 - Generating ERs for chr20"
#> [1] "2021-08-24 21:50:49 - Generating ERs for chr21"
#> [1] "2021-08-24 21:50:50 - Generating ERs for chr22"
#> [1] "2021-08-24 21:50:51 - Generating ERs for chrX"
#> [1] "2021-08-24 21:50:52 - Generating ERs for chrY"
#> [1] "2021-08-24 21:50:52 - Generating ERs for chrM"
#> [1] "2021-08-24 21:50:54 - Loading in GTF..."
#> [1] "2021-08-24 21:51:28 - Obtaining non-overlapping exons"
#> [1] "2021-08-24 21:51:29 - Calculating delta for ERs..."
#> [1] "2021-08-24 21:51:39 - Obtaining optimal set of ERs..."

opt_ers

#> $opt_ers
#> GRanges object with 345380 ranges and 0 metadata columns:
#>            seqnames            ranges strand
#>               <Rle>         <IRanges>  <Rle>
#>        [1]     chr1       14468-14829      +
#>        [2]     chr1       14970-14976      +
#>        [3]     chr1       15029-15038      +
#>        [4]     chr1       15796-15809      +
#>        [5]     chr1       15934-15947      +
#>        ...      ...               ...    ...
#>   [345376]     chrY 10192024-10192091      -
#>   [345377]     chrY 13286661-13286767      -
#>   [345378]     chrY 13286878-13286959      -
#>   [345379]     chrY 13287222-13287350      -
#>   [345380]     chrM           1-16569      -
#>   -------
#>   seqinfo: 25 sequences from an unspecified genome; no seqlengths
#> 
#> $opt_mcc_mrg
#> [1] "mcc_10" "mrg_20"
#> 
#> $deltas
#> # A tibble: 4 x 7
#>     mcc   mrg      sum  mean median n_eq_0 propor_eq_0
#>   <dbl> <dbl>    <int> <dbl>  <dbl>  <int>       <dbl>
#> 1     5    10 42342549  796.     79  19014       0.357
#> 2     5    20 33095527  686.     39  19267       0.399
#> 3    10    10 33189459  801.     60  16668       0.402
#> 4    10    20 25021958  667.     16  16897       0.450
#>