Finds overlapping junctions and annotates ERs as exonic, intronic, intergenic, some combination or none of those

Looks at junctions passed in to find any overlaps and adds them in along with other information as metadata columns. Then uses a gtf file or a Txdb passed in to generate a genomic state and then labels each ER as to whether they are exonic, intronic, intergenic on none

annotatERs(
  opt_ers,
  junc_data,
  gtf_path,
  txdb = NULL,
  chrs_to_keep = c(1:22, "X", "Y", "MT"),
  ensembl = TRUE
)

Arguments

opt_ers
optimally defined ERs (the product of the ODER function)

junc_data
junction data that should match the ERs passed into opt_ers

gtf_path
a gtf file with exon data

txdb
txdb object, if one is not entered a gtf file needs to be

chrs_to_keep
chromosomes to keep in genomic state (in NCBI format i.e. 1,2,3...22,X,Y,MT)

ensembl
logical variable to say whether the gtf file entered is from ensembl, the default is true

Value

annotated ERs

Examples

# \dontshow{
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[1])
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)
# }

opt_ers <- ODER(
    bw_paths = bw_path, auc_raw = auc_example,
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(5, 10), mrgs = c(10, 20),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
)

#> [1] "2021-08-24 21:59:02 - Obtaining mean coverage across 1 samples"
#> [1] "2021-08-24 21:59:02 - chr21"
#> [1] "2021-08-24 21:59:02 - chr22"
#> [1] "2021-08-24 21:59:03 - Generating ERs for chr21"
#> [1] "2021-08-24 21:59:04 - Generating ERs for chr22"
#> [1] "2021-08-24 21:59:05 - Loading in GTF..."
#> [1] "2021-08-24 21:59:39 - Obtaining non-overlapping exons"
#> [1] "2021-08-24 21:59:41 - Calculating delta for ERs..."
#> [1] "2021-08-24 21:59:41 - Obtaining optimal set of ERs..."

junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)
annot_ers <- annotatERs(
    opt_ers = opt_ers[["opt_ers"]], junc_data = junctions,
    gtf_path = gtf_path, chrs_to_keep = c("21", "22"), ensembl = TRUE
)

#> [1] "2021-08-24 21:59:41 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2021-08-24 21:59:41 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> [1] "2021-08-24 22:00:58 - Getting junction annotation using overlapping exons..."
#> [1] "2021-08-24 22:01:00 - Tidying junction annotation..."
#> [1] "2021-08-24 22:01:00 - Deriving junction categories..."
#> [1] "2021-08-24 22:01:05 - done!"
#> [1] "2021-08-24 22:01:05 - Finding junctions overlapping ers..."
#> [1] "2021-08-24 22:02:37 - Generating a genomic state..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens

#> 'select()' returned 1:1 mapping between keys and columns

#> [1] "2021-08-24 22:04:31 - Annotating the Expressed regions..."

#> 2021-08-24 22:04:31 annotateRegions: counting

#> 2021-08-24 22:04:31 annotateRegions: annotating

#> [1] "2021-08-24 22:05:28 - done!"

annot_ers

#> GRanges object with 15977 ranges and 5 metadata columns:
#>           seqnames            ranges strand |
#>              <Rle>         <IRanges>  <Rle> |
#>       [1]    chr21   5032176-5032217      * |
#>       [2]    chr21   5033408-5033425      * |
#>       [3]    chr21   5034717-5034756      * |
#>       [4]    chr21   5035188-5035189      * |
#>       [5]    chr21   5036577-5036581      * |
#>       ...      ...               ...    ... .
#>   [15973]    chr22 50798996-50799149      * |
#>   [15974]    chr22 50799209-50799284      * |
#>   [15975]    chr22 50799669-50799688      * |
#>   [15976]    chr22 50799717-50799744      * |
#>   [15977]    chr22 50800460-50800587      * |
#>                                                                                         grl
#>                                                                               <GRangesList>
#>       [1]                                                           chr21:5026423-5323718:+
#>       [2]                                                           chr21:5026423-5323718:+
#>       [3]                                                           chr21:5026423-5323718:+
#>       [4]                                                           chr21:5026423-5323718:+
#>       [5]                                                           chr21:5026423-5323718:+
#>       ...                                                                               ...
#>   [15973]     chr22:50363416-50804892:-,chr22:50460674-50807396:+,chr22:50770378-50805745:-
#>   [15974]     chr22:50363416-50804892:-,chr22:50460674-50807396:+,chr22:50770378-50805745:-
#>   [15975] chr22:50363416-50804892:-,chr22:50460674-50807396:+,chr22:50770378-50805745:-,...
#>   [15976] chr22:50363416-50804892:-,chr22:50460674-50807396:+,chr22:50770378-50805745:-,...
#>   [15977] chr22:50363416-50804892:-,chr22:50460674-50807396:+,chr22:50770378-50805745:-,...
#>                     genes  annotation  og_index       gene_source
#>           <CharacterList> <character> <integer>       <character>
#>       [1] ENSG00000277117        exon         1 nearest gtf genes
#>       [2] ENSG00000277117        exon         2 nearest gtf genes
#>       [3] ENSG00000277117        exon         3 nearest gtf genes
#>       [4] ENSG00000277117        exon         4 nearest gtf genes
#>       [5] ENSG00000277117        exon         5 nearest gtf genes
#>       ...             ...         ...       ...               ...
#>   [15973] ENSG00000184319        exon     15973 nearest gtf genes
#>   [15974] ENSG00000184319        exon     15974 nearest gtf genes
#>   [15975] ENSG00000184319        exon     15975 nearest gtf genes
#>   [15976] ENSG00000184319        exon     15976 nearest gtf genes
#>   [15977] ENSG00000184319        exon     15977 nearest gtf genes
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths