Generate the count matrix

Scores the mean coverage of the expressed regions as a count matrix

get_count_matrix(bw_paths, annot_ers, cols = NULL)

Arguments

bw_paths
Vector containing the bigwig file paths to read in

annot_ers
GRangesList containing the annotated ERs (product of annotatERs())

cols
A dataframe containing the information to be used as colData for the output. If NULL then the bw_paths will be used for the colData

Value

A Ranged Summarized Experiment containing the gene counts as an assay

Examples

# \dontshow{
library(magrittr)
gtex_metadata <- recount::all_metadata("gtex")

#> 2021-08-24 17:24:32 downloading the metadata to /tmp/RtmpKiYpcx/metadata_clean_gtex.Rdata

gtex_metadata <- gtex_metadata %>%
    as.data.frame() %>%
    dplyr::filter(project == "SRP012682")

gtf_url <- "http://ftp.ensembl.org/pub/release-103/gtf/homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
gtf_path <- ODER:::.file_cache(gtf_url)

url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[84])
bw_path2 <- ODER:::.file_cache(url[91])

opt_ers <- suppressWarnings(ODER(
    bw_paths = bw_path, auc_raw = gtex_metadata[["auc"]][84],
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(2, 4, 6, 8, 10), mrgs = c(10, 20, 30),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
))

#> [1] "2021-08-24 17:24:34 - Obtaining mean coverage across 1 samples"
#> [1] "2021-08-24 17:24:34 - chr21"
#> [1] "2021-08-24 17:24:34 - chr22"
#> [1] "2021-08-24 17:24:34 - Generating ERs for chr21"
#> [1] "2021-08-24 17:24:37 - Generating ERs for chr22"
#> [1] "2021-08-24 17:24:39 - Loading in GTF..."
#> [1] "2021-08-24 17:25:15 - Obtaining non-overlapping exons"
#> [1] "2021-08-24 17:25:19 - Calculating delta for ERs..."
#> [1] "2021-08-24 17:25:20 - Obtaining optimal set of ERs..."

opt_ers2 <- suppressWarnings(ODER(
    bw_paths = bw_path, auc_raw = gtex_metadata[["auc"]][91],
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(2, 4, 6, 8, 10), mrgs = c(10, 20, 30),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
))

#> [1] "2021-08-24 17:25:20 - Obtaining mean coverage across 1 samples"
#> [1] "2021-08-24 17:25:20 - chr21"
#> [1] "2021-08-24 17:25:21 - chr22"
#> [1] "2021-08-24 17:25:21 - Generating ERs for chr21"
#> [1] "2021-08-24 17:25:24 - Generating ERs for chr22"
#> [1] "2021-08-24 17:25:27 - Loading in GTF..."
#> [1] "2021-08-24 17:26:03 - Obtaining non-overlapping exons"
#> [1] "2021-08-24 17:26:05 - Calculating delta for ERs..."
#> [1] "2021-08-24 17:26:06 - Obtaining optimal set of ERs..."

test_juncs <- SummarizedExperiment::rowRanges(dasper::junctions_example)
aers <- suppressWarnings(annotatERs(opt_ers[["opt_ers"]],
    junc_data = test_juncs,
    gtf_path = gtf_path, ensembl = TRUE
))

#> [1] "2021-08-24 17:26:06 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2021-08-24 17:26:06 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> [1] "2021-08-24 17:27:26 - Getting junction annotation using overlapping exons..."
#> [1] "2021-08-24 17:27:28 - Tidying junction annotation..."
#> [1] "2021-08-24 17:27:28 - Deriving junction categories..."
#> [1] "2021-08-24 17:27:32 - done!"
#> [1] "2021-08-24 17:27:32 - Finding junctions overlapping ers..."
#> [1] "2021-08-24 17:27:58 - Generating a genomic state..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens

#> 'select()' returned 1:1 mapping between keys and columns

#> [1] "2021-08-24 17:34:23 - Annotating the Expressed regions..."

#> 2021-08-24 17:34:24 annotateRegions: counting

#> 2021-08-24 17:34:24 annotateRegions: annotating

#> [1] "2021-08-24 17:35:06 - done!"

aers2 <- suppressWarnings(annotatERs(opt_ers2[["opt_ers"]],
    junc_data = test_juncs,
    gtf_path = gtf_path, ensembl = TRUE
))

#> [1] "2021-08-24 17:35:06 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2021-08-24 17:35:06 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> [1] "2021-08-24 17:36:22 - Getting junction annotation using overlapping exons..."
#> [1] "2021-08-24 17:36:24 - Tidying junction annotation..."
#> [1] "2021-08-24 17:36:24 - Deriving junction categories..."
#> [1] "2021-08-24 17:36:28 - done!"
#> [1] "2021-08-24 17:36:28 - Finding junctions overlapping ers..."
#> [1] "2021-08-24 17:36:52 - Generating a genomic state..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens

#> 'select()' returned 1:1 mapping between keys and columns

#> [1] "2021-08-24 17:43:08 - Annotating the Expressed regions..."

#> 2021-08-24 17:43:09 annotateRegions: counting

#> 2021-08-24 17:43:09 annotateRegions: annotating

#> [1] "2021-08-24 17:43:52 - done!"

# }
example_bw_paths <- c(bw_path, bw_path2)
annot_ersl <- GenomicRanges::GRangesList(aers, aers2)

example_cm <- get_count_matrix(bw_paths = example_bw_paths, annot_ers = annot_ersl)

#> The latest megadepth version is 1.1.0c

#> megadepth has been installed to /github/home/bin

#> Warning: is.na() applied to non-(list or vector) of type 'S4'

#> Warning: is.na() applied to non-(list or vector) of type 'S4'

example_cm

#> class: RangedSummarizedExperiment 
#> dim: 5716 2 
#> metadata(0):
#> assays(1): ''
#> rownames: NULL
#> rowData names(0):
#> colnames(2): BFC5 BFC6
#> colData names(1): bw_paths