Generating the mean coverage of the expressed regions

get_coverage returns the mean coverage of the RNA sequencing data passed in. The data should be passed in via the file path to a Bigwig file.

get_coverage(
  bw_paths,
  auc_raw,
  auc_target,
  chrs = "",
  genome = "hg38",
  bw_chr = "chr"
)

Arguments

bw_paths
paths to bigwig files with the RNA-seq data that you want the coverage of.

auc_raw
vector containing AUCs(Area Under Coverage) matching the order of bigwig paths.

auc_target
total AUC to normalise all samples to. E.g. 40e6 * 100 would be the estimated total auc for sample sequenced to 40 million reads of 100bp in length.

chrs
chromosomes to obtain mean coverage for, default is "" giving every chromosome. Can take UCSC format(chrs = "chr1") or just the chromosome i.e. chrs = c(1,X)

genome
the UCSC genome you want to use, the default is hg38.

bw_chr
specifies whether the bigwig files has the chromosomes labelled with a "chr" preceding the chromosome i.e. "chr1" vs "1". Can be either "chr" or "nochr" with "chr" being the default.

Value

a list of Rles detailing the mean coverage per chromosome passed in.

Examples

# \dontshow{
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[1])
# }
eg_coverage <- get_coverage(
    bw_paths = bw_path,
    auc_raw = 11872688252,
    auc_target = 40e6 * 100,
    chrs = c("chr21", "chr22")
)

#> [1] "2021-05-17 16:47:52 - Obtaining mean coverage across 1 samples"
#> [1] "2021-05-17 16:47:52 - chr21"
#> [1] "2021-05-17 16:47:52 - chr22"

print(eg_coverage)

#> $chr21
#> $chr21$meanCoverage
#> numeric-Rle of length 46709983 with 351770 runs
#>   Lengths:  5010597       76      112       74 ...     2074       36    10111
#>   Values : 0.000000 0.336908 0.000000 0.336908 ... 0.000000 0.336908 0.000000
#> 
#> 
#> $chr22
#> $chr22$meanCoverage
#> numeric-Rle of length 50818468 with 619684 runs
#>   Lengths: 10519675        2        1        2 ...      160       76    13858
#>   Values : 0.000000 0.336908 0.673815 6.401246 ... 0.000000 0.336908 0.000000
#> 
#>