Define sets of ERs

get_ers defines expressed regions across an inputted range of mean coverage cut-offs (MCCs) and max region gaps (MRGs) from the coverage.

get_ers(coverage, mccs, mrgs)

Arguments

coverage
the coverage of the bigwig files passed into get_coverage.

mccs
numeric vector containing the mean coverage cut-offs to apply. MCCs are the mininmum number of reads that a base has to have to be considered as expressed.

mrgs
numeric vector containing the max region gaps to apply. MRGs are the maximum number of bases between ERs that can be below the MCC and the adjacent ERs will be merged.

Value

list containing sets of ERs, each generated using a particular combination of MCC and MRG.

Examples

# gtex_SRP012682_SRX222703_lung_coverage_1 is from the package data folder
eg_ers <- get_ers(
    coverage = gtex_SRP012682_SRX222703_lung_coverage_1,
    mccs = c(5, 10),
    mrgs = c(10, 20)
)
#> 2021-09-29 14:19:16 - Generating ERs for chr21
#> 2021-09-29 14:19:17 - Generating ERs for chr22

eg_ers
#> $mcc_5
#> $mcc_5$mrg_10
#> GRanges object with 26459 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]    chr21   5026470-5026525      *
#>       [2]    chr21   5027955-5028030      *
#>       [3]    chr21   5032160-5032217      *
#>       [4]    chr21   5033408-5033441      *
#>       [5]    chr21   5034694-5034787      *
#>       ...      ...               ...    ...
#>   [26455]    chr22 50799669-50799744      *
#>   [26456]    chr22 50799915-50799990      *
#>   [26457]    chr22 50800460-50800587      *
#>   [26458]    chr22 50802354-50802398      *
#>   [26459]    chr22 50802615-50802631      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $mcc_5$mrg_20
#> GRanges object with 24404 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]    chr21   5026470-5026525      *
#>       [2]    chr21   5027955-5028030      *
#>       [3]    chr21   5032160-5032217      *
#>       [4]    chr21   5033408-5033441      *
#>       [5]    chr21   5034694-5034787      *
#>       ...      ...               ...    ...
#>   [24400]    chr22 50799669-50799744      *
#>   [24401]    chr22 50799915-50799990      *
#>   [24402]    chr22 50800460-50800587      *
#>   [24403]    chr22 50802354-50802398      *
#>   [24404]    chr22 50802615-50802631      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> 
#> $mcc_10
#> $mcc_10$mrg_10
#> GRanges object with 17761 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]    chr21   5032176-5032217      *
#>       [2]    chr21   5033408-5033425      *
#>       [3]    chr21   5034717-5034756      *
#>       [4]    chr21   5035188-5035189      *
#>       [5]    chr21   5036577-5036581      *
#>       ...      ...               ...    ...
#>   [17757]    chr22 50799209-50799264      *
#>   [17758]    chr22 50799283-50799284      *
#>   [17759]    chr22 50799669-50799688      *
#>   [17760]    chr22 50799717-50799744      *
#>   [17761]    chr22 50800460-50800587      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $mcc_10$mrg_20
#> GRanges object with 15977 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]    chr21   5032176-5032217      *
#>       [2]    chr21   5033408-5033425      *
#>       [3]    chr21   5034717-5034756      *
#>       [4]    chr21   5035188-5035189      *
#>       [5]    chr21   5036577-5036581      *
#>       ...      ...               ...    ...
#>   [15973]    chr22 50798996-50799149      *
#>   [15974]    chr22 50799209-50799284      *
#>   [15975]    chr22 50799669-50799688      *
#>   [15976]    chr22 50799717-50799744      *
#>   [15977]    chr22 50800460-50800587      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#>