Calculates delta for sets of ERs

Calculates the median exon delta and the number of ERs with an exon delta of 0 by comparing each combination of MCC and MRG with the optimum exons from the ensembl database.

get_ers_delta(ers, opt_exons, delta_fun = ODER:::.delta)

Arguments

ers
Sets of ERs across various MCCs/MRGs - output of get_ers.

opt_exons
GRanges object that contains the regions that ideally, you want the ER definitions to match - output of get_exons.

delta_fun
Function that calculates the delta between ERs and opt_exons. Takes as input a set of ERs from ers and opt_exons. Then outputs a tibble/dataframe containing the summarised delta scores for that set of one set of ERs.

Value

tibble/dataframe containing summarised delta values. One row per set of ERs.

Examples

# \dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    # file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
    gtf_path <- file_cache(gtf_url)
}
if (!exists("eg_opt_exons")) {
    eg_opt_exons <- get_exons(
        gtf = gtf_path,
        ucsc_chr = TRUE,
        ignore.strand = TRUE
    )
}
#> [1] "2021-09-17 11:23:06 - Loading in GTF..."
#> [1] "2021-09-17 11:23:48 - Obtaining non-overlapping exons"
# }
if (!exists("eg_ers_delta")) {
    eg_ers_delta <- get_ers_delta(
        ers = ers_example, # ers_example is from the package data folder
        opt_exons = eg_opt_exons,
        delta_fun = ODER:::.delta
    ) # .delta is ODER's default, you can pass in your own if you have one
}
#> [1] "2021-09-17 11:23:52 - Calculating delta for ERs..."
#> Warning: numerical expression has 2 elements: only the first used
#> Warning: numerical expression has 2 elements: only the first used
print(eg_ers_delta)
#> # A tibble: 1 × 7
#>     mcc   mrg     sum  mean median n_eq_0 propor_eq_0
#>   <dbl> <dbl>   <int> <dbl>  <dbl>  <int>       <dbl>
#> 1     5    10 2187300  964.   150.    322       0.142