Get the expressed genes

Takes in a gtf file and a dataframe with expressed genes from a specific tissue. Filters the gtf file for genes and keeps those that are present in the tissue dataframe passed in.

get_expressed_genes(gtf_path, species = "Homo_sapiens", tissue_df)

Arguments

gtf_path
gtf file path

species
character string containing the species to filter for, Homo sapiens is the default

tissue_df
dataframe containing the expressed genes for a particular tissue

Value

GRanges with the expressed genes for a specific tissue

Examples

# \dontshow{
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
# .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
gtf_path <- ODER:::.file_cache(gtf_url)
# }
lung_tissue <- get_tissue(tissue = "lung")

lung_exp_genes <- get_expressed_genes(
    gtf_path = gtf_path,
    tissue_df = lung_tissue
)

lung_exp_genes

#> GRanges object with 21393 ranges and 22 metadata columns:
#>           seqnames        ranges strand |         source     type     score
#>              <Rle>     <IRanges>  <Rle> |       <factor> <factor> <numeric>
#>       [1]     chr1   14404-29570      - | havana             gene        NA
#>       [2]     chr1  89295-133723      - | ensembl_havana     gene        NA
#>       [3]     chr1 135141-135895      - | havana             gene        NA
#>       [4]     chr1 139790-140339      - | havana             gene        NA
#>       [5]     chr1 141474-173862      - | havana             gene        NA
#>       ...      ...           ...    ... .            ...      ...       ...
#>   [21389]     chrM   10760-12137      + |          insdc     gene        NA
#>   [21390]     chrM   12337-14148      + |          insdc     gene        NA
#>   [21391]     chrM   14149-14673      - |          insdc     gene        NA
#>   [21392]     chrM   14674-14742      - |          insdc     gene        NA
#>   [21393]     chrM   14747-15887      + |          insdc     gene        NA
#>               phase         gene_id gene_version   gene_name    gene_source
#>           <integer>     <character>  <character> <character>    <character>
#>       [1]      <NA> ENSG00000227232            5      WASH7P         havana
#>       [2]      <NA> ENSG00000238009            6  AL627309.1 ensembl_havana
#>       [3]      <NA> ENSG00000268903            1  AL627309.6         havana
#>       [4]      <NA> ENSG00000239906            1  AL627309.2         havana
#>       [5]      <NA> ENSG00000241860            7  AL627309.5         havana
#>       ...       ...             ...          ...         ...            ...
#>   [21389]      <NA> ENSG00000198886            2      MT-ND4          insdc
#>   [21390]      <NA> ENSG00000198786            2      MT-ND5          insdc
#>   [21391]      <NA> ENSG00000198695            2      MT-ND6          insdc
#>   [21392]      <NA> ENSG00000210194            1       MT-TE          insdc
#>   [21393]      <NA> ENSG00000198727            2      MT-CYB          insdc
#>                     gene_biotype transcript_id transcript_version
#>                      <character>   <character>        <character>
#>       [1] unprocessed_pseudogene          <NA>               <NA>
#>       [2]                 lncRNA          <NA>               <NA>
#>       [3]   processed_pseudogene          <NA>               <NA>
#>       [4]                 lncRNA          <NA>               <NA>
#>       [5]                 lncRNA          <NA>               <NA>
#>       ...                    ...           ...                ...
#>   [21389]         protein_coding          <NA>               <NA>
#>   [21390]         protein_coding          <NA>               <NA>
#>   [21391]         protein_coding          <NA>               <NA>
#>   [21392]                Mt_tRNA          <NA>               <NA>
#>   [21393]         protein_coding          <NA>               <NA>
#>           transcript_name transcript_source transcript_biotype         tag
#>               <character>       <character>        <character> <character>
#>       [1]            <NA>              <NA>               <NA>        <NA>
#>       [2]            <NA>              <NA>               <NA>        <NA>
#>       [3]            <NA>              <NA>               <NA>        <NA>
#>       [4]            <NA>              <NA>               <NA>        <NA>
#>       [5]            <NA>              <NA>               <NA>        <NA>
#>       ...             ...               ...                ...         ...
#>   [21389]            <NA>              <NA>               <NA>        <NA>
#>   [21390]            <NA>              <NA>               <NA>        <NA>
#>   [21391]            <NA>              <NA>               <NA>        <NA>
#>   [21392]            <NA>              <NA>               <NA>        <NA>
#>   [21393]            <NA>              <NA>               <NA>        <NA>
#>           transcript_support_level exon_number     exon_id exon_version
#>                        <character> <character> <character>  <character>
#>       [1]                     <NA>        <NA>        <NA>         <NA>
#>       [2]                     <NA>        <NA>        <NA>         <NA>
#>       [3]                     <NA>        <NA>        <NA>         <NA>
#>       [4]                     <NA>        <NA>        <NA>         <NA>
#>       [5]                     <NA>        <NA>        <NA>         <NA>
#>       ...                      ...         ...         ...          ...
#>   [21389]                     <NA>        <NA>        <NA>         <NA>
#>   [21390]                     <NA>        <NA>        <NA>         <NA>
#>   [21391]                     <NA>        <NA>        <NA>         <NA>
#>   [21392]                     <NA>        <NA>        <NA>         <NA>
#>   [21393]                     <NA>        <NA>        <NA>         <NA>
#>            protein_id protein_version     ccds_id
#>           <character>     <character> <character>
#>       [1]        <NA>            <NA>        <NA>
#>       [2]        <NA>            <NA>        <NA>
#>       [3]        <NA>            <NA>        <NA>
#>       [4]        <NA>            <NA>        <NA>
#>       [5]        <NA>            <NA>        <NA>
#>       ...         ...             ...         ...
#>   [21389]        <NA>            <NA>        <NA>
#>   [21390]        <NA>            <NA>        <NA>
#>   [21391]        <NA>            <NA>        <NA>
#>   [21392]        <NA>            <NA>        <NA>
#>   [21393]        <NA>            <NA>        <NA>
#>   -------
#>   seqinfo: 25 sequences from an unspecified genome; no seqlengths