Get junctions that overlap the optimally defined ERs

The junctions will be added to the ERs passed in as a metadata column, a GRangeslist with each GRanges element of the list corresponding to it's associated ER. If there are no overlaps, the GRangeslist will be empty. Any ERs that do not have an overlapping junction will have an empty GRanges. The respective genes that each ER could be associated with will also be passed in as a metadata column, a character list.

get_junctions(opt_ers, junc_data, txdb)

Arguments

opt_ers
optimally defined ERs (the product of the ODER function)

junc_data
junction data that should match the ERs passed into opt_ers

txdb
a gtf file or txdb (TxDb-class) with exon data. This will ultimately be used to annotate the junction data passed in.

Value

optimally defined ers annotated with junction and gene information

See also

Examples

gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
# file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
gtf_path <- file_cache(gtf_url)

example_ers <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr21", "chr22"), c(10, 1)),
    ranges = IRanges::IRanges(
        start = c(
            5026423, 24738, 5032218, 5033895, 17554, 50800446,
            50800539, 16570, 50800790, 15005, 20312
        ),
        end = c(
            5323718, 24891, 5033407, 5033980, 17728, 50800910,
            50800817, 16723, 50800910, 15038, 20582
        ),
        names = head(letters, 11)
    ),
    strand = S4Vectors::Rle((c("+", "-")), c(6, 5)),
    score = 1:11,
    GC = seq(1, 0, length = 11)
)
if (!exists("gtf_gr")) {
    gtf_gr <- rtracklayer::import(gtf_path)
}
chrs_to_keep <- c("21", "22")
#### preparing the txdb and genomstate object(s)
hg38_chrominfo <- GenomeInfoDb::getChromInfoFromUCSC("hg38")
new_info <- hg38_chrominfo$size[match(
    chrs_to_keep,
    GenomeInfoDb::mapSeqlevels(hg38_chrominfo$chrom, "Ensembl")
)]
names(new_info) <- chrs_to_keep
gtf_gr_tx <- GenomeInfoDb::keepSeqlevels(gtf_gr,
    chrs_to_keep,
    pruning.mode = "tidy"
)
GenomeInfoDb::seqlengths(gtf_gr_tx) <- new_info
GenomeInfoDb::seqlevelsStyle(gtf_gr_tx) <- "UCSC"
GenomeInfoDb::genome(gtf_gr_tx) <- "hg38"

ucsc_txdb <- GenomicFeatures::makeTxDbFromGRanges(gtf_gr_tx)
#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.
genom_state <- derfinder::makeGenomicState(txdb = ucsc_txdb)
#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens
#> 'select()' returned 1:1 mapping between keys and columns
ens_txdb <- ucsc_txdb
GenomeInfoDb::seqlevelsStyle(ens_txdb) <- "Ensembl"
################### end of genomstate creation

junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)

example_er_juncs <- get_junctions(
    opt_ers = example_ers,
    junc_data = junctions,
    txdb = ens_txdb # can either be a gtf file or txdb in the Ensembl format
)
#> [1] "2021-09-29 14:21:41 - Obtaining co-ordinates of annotated exons and junctions..."
#> [1] "2021-09-29 14:21:42 - Getting junction annotation using overlapping exons..."
#> [1] "2021-09-29 14:21:44 - Tidying junction annotation..."
#> [1] "2021-09-29 14:21:44 - Deriving junction categories..."
#> [1] "2021-09-29 14:21:49 - done!"
#> 2021-09-29 14:21:49 - Finding junctions overlapping ers...

example_er_juncs
#> GRanges object with 11 ranges and 4 metadata columns:
#>     seqnames            ranges strand |     score        GC
#>        <Rle>         <IRanges>  <Rle> | <integer> <numeric>
#>   a    chr21   5026423-5323718      + |         1       1.0
#>   b    chr21       24738-24891      + |         2       0.9
#>   c    chr21   5032218-5033407      + |         3       0.8
#>   d    chr21   5033895-5033980      + |         4       0.7
#>   e    chr21       17554-17728      + |         5       0.6
#>   f    chr21 50800446-50800910      + |         6       0.5
#>   g    chr21 50800539-50800817      - |         7       0.4
#>   h    chr21       16570-16723      - |         8       0.3
#>   i    chr21 50800790-50800910      - |         9       0.2
#>   j    chr21       15005-15038      - |        10       0.1
#>   k    chr22       20312-20582      - |        11       0.0
#>                                                                             grl
#>                                                                   <GRangesList>
#>   a chr21:5026423-5323718:+,chr21:5028226-5032052:+,chr21:5032218-5033407:+,...
#>   b                                                                            
#>   c                             chr21:5026423-5323718:+,chr21:5032218-5033407:+
#>   d                             chr21:5026423-5323718:+,chr21:5033895-5033980:+
#>   e                                                                            
#>   f                                                                            
#>   g                                                                            
#>   h                                                                            
#>   i                                                                            
#>   j                                                                            
#>   k                                                                            
#>               genes
#>     <CharacterList>
#>   a ENSG00000277117
#>   b                
#>   c ENSG00000277117
#>   d                
#>   e                
#>   f                
#>   g                
#>   h                
#>   i                
#>   j                
#>   k                
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths