Define sets of ERs for stranded Bigwigs

get_strand_ers defines ERs across an inputted range of mean coverage cut-offs (MCCs) and max region gaps (MRGs) from the coverage.

get_strand_ers(
  bw_pos,
  bw_neg,
  auc_raw_pos,
  auc_raw_neg,
  auc_target,
  chrs,
  mccs,
  mrgs,
  bw_chr = "chr"
)

Arguments

bw_pos
positive strand bigwig file

bw_neg
negative strand bigwig file

auc_raw_pos
vector containing AUCs(Area Under Coverage) matching the order of the positive bigwig paths.

auc_raw_neg
vector containing AUCs(Area Under Coverage) matching the order of the negative bigwig paths.

auc_target
total AUC to normalise all samples to. E.g. 40e6 * 100 would be the estimated total auc for sample sequenced to 40 million reads of 100bp in length.

chrs
chromosomes to obtain mean coverage for, default is "" giving every chromosome. Can take UCSC format(chrs = "chr1") or just the chromosome i.e. chrs = c(1,X)

mccs
mean coverage cut-offs to apply.

mrgs
max region gaps to apply.

bw_chr
specifies whether the bigwig files has the chromosomes labelled with a "chr" preceding the chromosome i.e. "chr1" vs "1". Can be either "chr" or "nochr" with "chr" being the default.

Value

list containing sets of stranded ERs, each generated using a particular combination of MCC and MRG.

Examples

library("magrittr")
if (!exists("gtex_metadata")) {
    gtex_metadata <- recount::all_metadata("gtex")
    gtex_metadata <- gtex_metadata %>%
        as.data.frame() %>%
        dplyr::filter(project == "SRP012682")
}
#> 2021-09-21 11:15:46 downloading the metadata to /tmp/RtmpiiwrQy/metadata_clean_gtex.Rdata
if (!exists("rec_url")) {
    rec_url <- recount::download_study(
        project = "SRP012682",
        type = "samples",
        download = FALSE
    )
}
# file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
bw_plus <- file_cache(rec_url[58])
bw_minus <- file_cache(rec_url[84])

# As of rtracklayer 1.25.16, BigWig is not supported on Windows.
if (!xfun::is_windows()) {
    stranded_ers <- get_strand_ers(
        bw_pos = bw_plus, bw_neg = bw_minus,
        auc_raw_pos = gtex_metadata[["auc"]][58],
        auc_raw_neg = gtex_metadata[["auc"]][84], auc_target = 40e6 * 100,
        chrs = "chr21", mccs = c(5, 10), mrgs = c(10, 20)
    )
    stranded_ers
}
#> 2021-09-21 11:15:48 - Obtaining mean coverage across 1 samples
#> 2021-09-21 11:15:48 - chr21
#> 2021-09-21 11:15:49 - Obtaining mean coverage across 1 samples
#> 2021-09-21 11:15:49 - chr21
#> 2021-09-21 11:15:49 - Generating ERs for chr21
#> 2021-09-21 11:15:50 - Generating ERs for chr21
#> $mcc_5
#> $mcc_5$mrg_10
#> GRanges object with 6355 ranges and 0 metadata columns:
#>          seqnames            ranges strand
#>             <Rle>         <IRanges>  <Rle>
#>      [1]    chr21   5116346-5117232      +
#>      [2]    chr21   5117340-5117507      +
#>      [3]    chr21   5117539-5117558      +
#>      [4]    chr21   5117587-5117592      +
#>      [5]    chr21   5117605-5117611      +
#>      ...      ...               ...    ...
#>   [6351]    chr21 46668354-46668391      -
#>   [6352]    chr21 46668488-46668581      -
#>   [6353]    chr21 46690842-46691238      -
#>   [6354]    chr21 46691775-46691807      -
#>   [6355]    chr21 46692377-46692446      -
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $mcc_5$mrg_20
#> GRanges object with 5503 ranges and 0 metadata columns:
#>          seqnames            ranges strand
#>             <Rle>         <IRanges>  <Rle>
#>      [1]    chr21   5116346-5117232      +
#>      [2]    chr21   5117340-5117507      +
#>      [3]    chr21   5117539-5117558      +
#>      [4]    chr21   5117587-5117611      +
#>      [5]    chr21   5117670-5117817      +
#>      ...      ...               ...    ...
#>   [5499]    chr21 46668354-46668391      -
#>   [5500]    chr21 46668488-46668581      -
#>   [5501]    chr21 46690842-46691238      -
#>   [5502]    chr21 46691775-46691807      -
#>   [5503]    chr21 46692377-46692446      -
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> 
#> $mcc_10
#> $mcc_10$mrg_10
#> GRanges object with 4293 ranges and 0 metadata columns:
#>          seqnames            ranges strand
#>             <Rle>         <IRanges>  <Rle>
#>      [1]    chr21   5116356-5117231      +
#>      [2]    chr21   5117368-5117396      +
#>      [3]    chr21   5118691-5118847      +
#>      [4]    chr21   5120950-5120951      +
#>      [5]    chr21   5121718-5121803      +
#>      ...      ...               ...    ...
#>   [4289]    chr21 46665571-46665576      -
#>   [4290]    chr21 46665603-46665610      -
#>   [4291]    chr21 46690860-46691105      -
#>   [4292]    chr21 46691173-46691177      -
#>   [4293]    chr21 46691195-46691200      -
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $mcc_10$mrg_20
#> GRanges object with 3814 ranges and 0 metadata columns:
#>          seqnames            ranges strand
#>             <Rle>         <IRanges>  <Rle>
#>      [1]    chr21   5116356-5117231      +
#>      [2]    chr21   5117368-5117396      +
#>      [3]    chr21   5118691-5118847      +
#>      [4]    chr21   5120950-5120951      +
#>      [5]    chr21   5121718-5121803      +
#>      ...      ...               ...    ...
#>   [3810]    chr21 46665544-46665545      -
#>   [3811]    chr21 46665571-46665576      -
#>   [3812]    chr21 46665603-46665610      -
#>   [3813]    chr21 46690860-46691105      -
#>   [3814]    chr21 46691173-46691200      -
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#>