Refines the ERs start and end points

Uses the junctions in the metadata to modify the starts and ends of the Expressed regions

refine_ERs(annot_ers)

Arguments

annot_ers
ERs that have been annotated (result of annotatER)

Value

Genomic ranges with refined base pair starts and ends and a logical vector listing changes

Examples

# \dontshow{
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[1])
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)
# }

opt_ers <- ODER(
    bw_paths = bw_path, auc_raw = auc_example,
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(5, 10), mrgs = c(10, 20),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
)

#> [1] "2021-07-08 15:14:59 - Obtaining mean coverage across 1 samples"
#> [1] "2021-07-08 15:14:59 - chr21"
#> [1] "2021-07-08 15:15:00 - chr22"
#> [1] "2021-07-08 15:15:00 - Generating ERs for chr21"
#> [1] "2021-07-08 15:15:01 - Generating ERs for chr22"
#> [1] "2021-07-08 15:15:02 - Loading in GTF..."
#> [1] "2021-07-08 15:15:34 - Obtaining non-overlapping exons"
#> [1] "2021-07-08 15:15:36 - Calculating delta for ERs..."
#> [1] "2021-07-08 15:15:36 - Obtaining optimal set of ERs..."

junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)
annot_ers <- annotatERs(
    opt_ers = opt_ers[["opt_ers"]], junc_data = junctions,
    gtf_path = gtf_path, chrs_to_keep = c("21", "22"), ensembl = TRUE
)

#> [1] "2021-07-08 15:15:36 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2021-07-08 15:15:36 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> [1] "2021-07-08 15:16:53 - Getting junction annotation using overlapping exons..."
#> [1] "2021-07-08 15:16:55 - Tidying junction annotation..."
#> [1] "2021-07-08 15:16:55 - Deriving junction categories..."
#> [1] "2021-07-08 15:16:59 - done!"
#> [1] "2021-07-08 15:16:59 - Finding junctions overlapping ers..."
#> [1] "2021-07-08 15:18:20 - Generating a genomic state..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens

#> 'select()' returned 1:1 mapping between keys and columns

#> [1] "2021-07-08 15:20:11 - Annotating the Expressed regions..."

#> 2021-07-08 15:20:11 annotateRegions: counting

#> 2021-07-08 15:20:11 annotateRegions: annotating

#> [1] "2021-07-08 15:21:07 - done!"

refined_ers <- refine_ERs(annot_ers)

#> [1] "2021-07-08 15:21:07 - Refining the Expressed regions..."

refined_ers

#> GRanges object with 0 ranges and 5 metadata columns:
#>    seqnames    ranges strand |           grl           genes  annotation
#>       <Rle> <IRanges>  <Rle> | <GRangesList> <CharacterList> <character>
#>     og_index gene_source
#>    <integer> <character>
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths