Define sets of ERs

gen_ERs defines ERs across an inputted range of mean coverage cut-offs (MCCs) and max region gaps (MRGs) from the coverage.

get_ers(coverage, mccs, mrgs)

Arguments

coverage
the coverage of the bigwig files passed into get_coverage.

mccs
mean coverage cut-offs to apply.

mrgs
max region gaps to apply.

Value

list containing sets of ERs, each generated using a particular combination of MCC and MRG.

Examples

# coverage_example is from the package data folder
eg_ers <- get_ers(coverage = coverage_example, mccs = c(5, 10), mrgs = c(10, 20))

#> [1] "2021-05-12 10:17:09 - Generating ERs for chr21"
#> [1] "2021-05-12 10:17:10 - Generating ERs for chr22"

print(eg_ers)

#> $mcc_5
#> $mcc_5$mrg_10
#> GRanges object with 26459 ranges and 0 metadata columns:
#>         seqnames            ranges strand
#>            <Rle>         <IRanges>  <Rle>
#>   chr21    chr21   5026470-5026525      *
#>   chr21    chr21   5027955-5028030      *
#>   chr21    chr21   5032160-5032217      *
#>   chr21    chr21   5033408-5033441      *
#>   chr21    chr21   5034694-5034787      *
#>     ...      ...               ...    ...
#>   chr22    chr22 50799669-50799744      *
#>   chr22    chr22 50799915-50799990      *
#>   chr22    chr22 50800460-50800587      *
#>   chr22    chr22 50802354-50802398      *
#>   chr22    chr22 50802615-50802631      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $mcc_5$mrg_20
#> GRanges object with 24404 ranges and 0 metadata columns:
#>         seqnames            ranges strand
#>            <Rle>         <IRanges>  <Rle>
#>   chr21    chr21   5026470-5026525      *
#>   chr21    chr21   5027955-5028030      *
#>   chr21    chr21   5032160-5032217      *
#>   chr21    chr21   5033408-5033441      *
#>   chr21    chr21   5034694-5034787      *
#>     ...      ...               ...    ...
#>   chr22    chr22 50799669-50799744      *
#>   chr22    chr22 50799915-50799990      *
#>   chr22    chr22 50800460-50800587      *
#>   chr22    chr22 50802354-50802398      *
#>   chr22    chr22 50802615-50802631      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> 
#> $mcc_10
#> $mcc_10$mrg_10
#> GRanges object with 17761 ranges and 0 metadata columns:
#>         seqnames            ranges strand
#>            <Rle>         <IRanges>  <Rle>
#>   chr21    chr21   5032176-5032217      *
#>   chr21    chr21   5033408-5033425      *
#>   chr21    chr21   5034717-5034756      *
#>   chr21    chr21   5035188-5035189      *
#>   chr21    chr21   5036577-5036581      *
#>     ...      ...               ...    ...
#>   chr22    chr22 50799209-50799264      *
#>   chr22    chr22 50799283-50799284      *
#>   chr22    chr22 50799669-50799688      *
#>   chr22    chr22 50799717-50799744      *
#>   chr22    chr22 50800460-50800587      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $mcc_10$mrg_20
#> GRanges object with 15977 ranges and 0 metadata columns:
#>         seqnames            ranges strand
#>            <Rle>         <IRanges>  <Rle>
#>   chr21    chr21   5032176-5032217      *
#>   chr21    chr21   5033408-5033425      *
#>   chr21    chr21   5034717-5034756      *
#>   chr21    chr21   5035188-5035189      *
#>   chr21    chr21   5036577-5036581      *
#>     ...      ...               ...    ...
#>   chr22    chr22 50798996-50799149      *
#>   chr22    chr22 50799209-50799284      *
#>   chr22    chr22 50799669-50799688      *
#>   chr22    chr22 50799717-50799744      *
#>   chr22    chr22 50800460-50800587      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#>