Get junctions that overlap the optimally defined ERs

The junctions will be added to the ERs passed in as a metadata column, a GRangeslist with each GRanges element of the list corresponding to it's associated ER. Any ERs that do not have an overlapping junction will have an empty GRanges. The respective genes that each ER could be associated with will also be passed in as a metadata column, a character list.

get_junctions(opt_ers, junc_data, gtf_path)

Arguments

opt_ers
optimally defined ERs (the product of the ODER function)

junc_data
junction data that should match the ERs passed into opt_ers

gtf_path
a gtf file with exon data

Value

optimally defined ers annotated with junction and gene information

Details

Each Granges of the GRangeslist will have the metadata columns of "in_ref", "gene_id_start", "tx_name_start", "exon_name_start", "strand_start", "exon_width_start","gene_id_end", "tx_name_end", "exon_name_end", "strand_end", "exon_width_end", "gene_id_junction", "strand_junction", "type" and "er_index add the columns I'm adding

See also

Examples

# \dontshow{
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
# .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
gtf_path <- ODER:::.file_cache(gtf_url)
# }

example_ers <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr21", "chr22"), c(10, 1)),
    ranges = IRanges::IRanges(
        start = c(
            5026423, 24738, 5032218, 5033895, 17554, 50800446,
            50800539, 16570, 50800790, 15005, 20312
        ),
        end = c(
            5323718, 24891, 5033407, 5033980, 17728, 50800910,
            50800817, 16723, 50800910, 15038, 20582
        ),
        names = head(letters, 11)
    ),
    strand = S4Vectors::Rle((c("+", "-")), c(6, 5)),
    score = 1:11,
    GC = seq(1, 0, length = 11)
)

example_junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)

example_er_juncs <- get_junctions(
    opt_ers = example_ers,
    junc_data = example_junctions,
    gtf_path = gtf_path
)

#> [1] "2021-07-25 16:27:06 - Obtaining co-ordinates of annotated exons and junctions from gtf/gff3..."
#> [1] "2021-07-25 16:27:06 - Importing gtf/gff3 as a TxDb..."

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> Warning: The "phase" metadata column contains non-NA values for features of type
#>   stop_codon. This information was ignored.

#> OK

#> [1] "2021-07-25 16:28:20 - Getting junction annotation using overlapping exons..."
#> [1] "2021-07-25 16:28:22 - Tidying junction annotation..."
#> [1] "2021-07-25 16:28:22 - Deriving junction categories..."
#> [1] "2021-07-25 16:28:25 - done!"
#> [1] "2021-07-25 16:28:25 - Finding junctions overlapping ers..."

print(example_er_juncs)

#> GRanges object with 11 ranges and 4 metadata columns:
#>     seqnames            ranges strand |     score        GC
#>        <Rle>         <IRanges>  <Rle> | <integer> <numeric>
#>   a    chr21   5026423-5323718      + |         1       1.0
#>   b    chr21       24738-24891      + |         2       0.9
#>   c    chr21   5032218-5033407      + |         3       0.8
#>   d    chr21   5033895-5033980      + |         4       0.7
#>   e    chr21       17554-17728      + |         5       0.6
#>   f    chr21 50800446-50800910      + |         6       0.5
#>   g    chr21 50800539-50800817      - |         7       0.4
#>   h    chr21       16570-16723      - |         8       0.3
#>   i    chr21 50800790-50800910      - |         9       0.2
#>   j    chr21       15005-15038      - |        10       0.1
#>   k    chr22       20312-20582      - |        11       0.0
#>                                                                             grl
#>                                                                   <GRangesList>
#>   a chr21:5026423-5323718:+,chr21:5028226-5032052:+,chr21:5032218-5033407:+,...
#>   b                                                                            
#>   c                             chr21:5026423-5323718:+,chr21:5032218-5033407:+
#>   d                             chr21:5026423-5323718:+,chr21:5033895-5033980:+
#>   e                                                                            
#>   f                                                                            
#>   g                                                                            
#>   h                                                                            
#>   i                                                                            
#>   j                                                                            
#>   k                                                                            
#>               genes
#>     <CharacterList>
#>   a ENSG00000277117
#>   b                
#>   c ENSG00000277117
#>   d                
#>   e                
#>   f                
#>   g                
#>   h                
#>   i                
#>   j                
#>   k                
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths