Obtains optimised set of ERs

Uses a delta calculating function and a well defined set of exons to find which combination of MCC and MRG gives the best definition of the Expressed regions.

get_opt_ers(ers, ers_delta)

Arguments

ers
Sets of ERs across various MCCs/MRGs - output of get_ers.

ers_delta
tibble/dataframe containing summarised delta values. One row per set of ERs.

Value

list containing optimised ERs, optimal pair of MCC/MRGs and delta_df

Examples

# \dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    gtf_path <- file_cache(gtf_url)
}
if (!exists("eg_opt_exons")) {
    eg_opt_exons <- get_exons(
        gtf = gtf_path,
        ucsc_chr = TRUE,
        ignore.strand = TRUE
    )
}
#> 2021-09-21 16:18:25 - Loading in GTF...
#> 2021-09-21 16:19:05 - Obtaining non-overlapping exons
if (!exists("eg_ers_delta")) {
    eg_ers_delta <- get_ers_delta(
        ers = gtex_lung_ers_1, # gtex_lung_ers_1 is from the package data folder
        opt_exons = eg_opt_exons
    )
}
#> 2021-09-21 16:19:08 - Calculating delta for ERs...
# }
opt_ers <- get_opt_ers(
    ers = gtex_lung_ers_1,
    ers_delta = eg_ers_delta
)
#> 2021-09-21 16:19:09 - Obtaining optimal set of ERs...
opt_ers
#> $opt_ers
#> GRanges object with 15977 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]    chr21   5032176-5032217      *
#>       [2]    chr21   5033408-5033425      *
#>       [3]    chr21   5034717-5034756      *
#>       [4]    chr21   5035188-5035189      *
#>       [5]    chr21   5036577-5036581      *
#>       ...      ...               ...    ...
#>   [15973]    chr22 50798996-50799149      *
#>   [15974]    chr22 50799209-50799284      *
#>   [15975]    chr22 50799669-50799688      *
#>   [15976]    chr22 50799717-50799744      *
#>   [15977]    chr22 50800460-50800587      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $opt_mcc_mrg
#> [1] "mcc_10" "mrg_20"
#> 
#> $deltas
#> # A tibble: 4 × 7
#>     mcc   mrg     sum  mean median n_eq_0 propor_eq_0
#>   <dbl> <dbl>   <int> <dbl>  <dbl>  <int>       <dbl>
#> 1     5    10 2187300  964.   150.    322       0.142
#> 2     5    20 1892386  898.   141     335       0.159
#> 3    10    10 1771837 1009.   139     316       0.180
#> 4    10    20 1457726  911.   118     323       0.202
#>